updating test structure
