+++ lib/msal-core/karma.conf.js
@@ -7,16 +7,23 @@ module.exports = function (config) {
         frameworks: ['mocha'],
         files: [
             'node_modules/babel-polyfill/dist/polyfill.js',
-            'tests/mochaTests/**/*.ts',
+            'tests/**/*.ts',
+        ],
+        exclude: [
+            'tests/MSAL*',
+            'tests/Authority.spec.ts'
         ],
         preprocessors: {
-            'tests/mochaTests/**/*.ts': ['webpack', 'sourcemap'],
+            'tests/**/*.ts': ['webpack', 'sourcemap'],
             'src/**/ *.ts': ['coverage']
         },
         webpack: {
             mode: webpackConfig.mode,
             module: webpackConfig.module,
-            resolve: webpackConfig.resolve
+            resolve: webpackConfig.resolve,
+            performance: {
+                hints: false
+            }
         },
         reporters: ['progress'],
         port: 8080,
