+++ lib/msal-core/tests/error/InteractionRequiredAuthError.spec.ts
@@ -1,6 +1,6 @@
 import * as mocha from "mocha";
 import * as chai from "chai";
-import { InteractionRequiredAuthError, InteractionRequiredAuthErrorMessage } from "../../../src/error/InteractionRequiredAuthError";
+import { InteractionRequiredAuthError, InteractionRequiredAuthErrorMessage } from "../../src/error/InteractionRequiredAuthError";
 
 const expect = chai.expect;
 chai.config.includeStack = false;
