+++ lib/msal-core/webpack.config.js
@@ -49,7 +49,6 @@ module.exports = {
           transpileOnly: true,
         }
       }
-
     }]
   }
 }
